Refactor to suit configurable diffing
